separated routing code into routers organized project folders leaderboard unranked visual fix
