+++ package.json
@@ -10,8 +10,8 @@
     "serve-favicon": "^2.5.0"
   },
   "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js"
+    "start": "node src/app.js",
+    "dev": "nodemon src/app.js"
   },
   "keywords": [
     "node",
