+++ views/pages/index.ejs
@@ -10,8 +10,6 @@
         <%- await include('../partials/header') %>
     </header>
     <main>
-        <%- await include('../partials/live-games', {liveGames: liveGames}) %>
-
         <div class="jumbotron">
             <h3>Groups</h3>
             <% groupNames.forEach((name, index) => { %>
