+++ views/partials/header.ejs
@@ -2,7 +2,7 @@
     <a class="navbar-brand" href="/">League Dev</a>
     <ul class="navbar-nav mr-auto">
         <li class="nav-item">
-            <!-- <a class="nav-link" href="/test">Test</a> -->
+            <a class="nav-link" href="/">Test</a>
         </li>
     </ul>
 </nav>
