+++ views/partials/liveGames.ejs
